Remove benchmarks that failed to compile from the generated config files
Previously this was only done for explicitly excluded benchmarks, such as
the Fortran benchmarks in case of clang. Also mention the name of the
config files from which they are removed.

+++ speccpu2006/spec2regression.sh
@@ -129,7 +129,7 @@ fi
 MISSING_BENCHMARKS=
 for dir in "$SPEC_INSTALLED_DIR"/benchspec/CPU2006/*/; do
   benchdir=`basename "$dir"`
-  if [ -d "$dir"/build/"$SPEC_BUILD_DIR" ]; then
+  if [ -d "$dir"/build/"$SPEC_BUILD_DIR" ] && ! grep -q "^specmake: .*Error" "$dir"/build/"$SPEC_BUILD_DIR"/make.err; then
     if [ "x${SPEC_COPY_BENCHMARKS}" = xy ]; then
       if [ -d "$TARGET_DIR"/"$benchdir" ] ; then
         rm -rf "$TARGET_DIR"/"$benchdir"
@@ -223,7 +223,7 @@ for conffile in *.conf; do
   for bench in $MISSING_BENCHMARKS; do
     benchlinestart=`grep -n $bench "$TARGET_DIR"/$conffile 2>/dev/null | head -n 1 | sed -e 's/:.*//'`
     if [ ! -z "$benchlinestart" ]; then
-      echo "Removing $bench from config file because it was not compiled (note that clang does not support Fortran benchmarks)"
+      echo "Removing $bench from "$TARGET_DIR"/$conffile because it was not (correctly) compiled"
       (
         head -n $(($benchlinestart-1)) < "$TARGET_DIR"/$conffile
         tail -n +$(($benchlinestart+5)) < "$TARGET_DIR"/$conffile
